+++ app/libs/php/actividades_model.php
@@ -76,15 +76,13 @@ class Actividades_Model extends Model{
         
         try
         {
-            $fecha_actual = date("Y-m-d");
-
-            $sql = "INSERT INTO actividades (idusuario,actividad,cliente,comentarios,fecha_inicio) VALUES ('{$actividad["idusuario"]}', '{$actividad["actividad"]}', '{$actividad["cliente"]}', '{$actividad["comentarios"]}','{$fecha_actual}')";
+            $sql = "INSERT INTO actividades (idusuario,actividad,comentarios,fecha_inicio,fecha_fin,precio_x_hora) VALUES ('{$actividad["idusuario"]}', '{$actividad["actividad"]}', '{$actividad["comentarios"]}', '{$actividad["fechaInicio"]}', '{$actividad["fechaFin"]}', '{$actividad["precio_x_hora"]}')";
             $result =  $this->db->query($sql);
-
-            
             $id = $this->db->insert_id;
-
-            return $id;
+            if (isset($id) && is_integer($id) && $id >0)
+                return $id;
+            else
+                return NULL;
         }
         catch(Exception $e){
             echo 'ERROR:'.$e->getMessage()."<br>";
@@ -93,6 +91,23 @@ class Actividades_Model extends Model{
             // //Manejo SIN excepciones (normal) de errores mysqli
             // if ($this->db->error)                        
             //     echo $this->db->error."<br>";
+        }
+    }
+
+    function editarActividad($actividad)
+    {
+        try
+        {
+            $sql = "UPDATE actividades SET actividad = '{$actividad["actividad"]}', comentarios = '{$actividad["comentarios"]}', precio_x_hora = '{$actividad["precio_x_hora"]}', fecha_inicio = '{$actividad["fecha_inicio"]}', fecha_fin = '{$actividad["fecha_fin"]}' WHERE idactividad = '{$actividad["idactividad"]}' AND idusuario = '{$actividad["idusuario"]}'";
+            
+            if ($this->db->query($sql))
+                return true;
+            else
+                return false;
+        }
+        catch(Exception $e){
+            echo 'ERROR:'.$e->getMessage()."<br>";
+            die();
         }
     }
 
@@ -105,7 +120,7 @@ class Actividades_Model extends Model{
                 while ($actividad = $query->fetch_object())
                     $actividades[] = $actividad;
         
-            if (count($actividades))
+            if (isset($actividades) && count($actividades)>0)
                 return $actividades;
             else
                 return 0;
@@ -124,7 +139,6 @@ class Actividades_Model extends Model{
         $actividad["cliente"] = isset($actividad['cliente']) ? mysqli::real_escape_string($actividad["cliente"]) : false;
         $actividad["comentarios"] = isset($actividad['comentarios']) ? mysqli::real_escape_string($actividad["comentarios"]) : false;
         $actividad["precio_x_hora"] = isset($actividad['precio_x_hora']) ? mysqli::real_escape_string($actividad["precio_x_hora"]) : false;
-        $actividad["precio_estimado"]  = isset($actividad['precio_estimado']) ? mysqli::real_escape_string($actividad["precio_estimado"]) : false;
 
             // Validación
             $respuesta['errores'] = array();
@@ -150,10 +164,6 @@ class Actividades_Model extends Model{
             if(empty($actividad['precio_x_hora']) || !is_float($actividad['precio_x_hora']))
                 $respuesta['errores']['precio_x_hora'] = 'Precio hora vacío o no es un número válido';
 
-            if(empty($actividad['estimado']) || !is_float($actividad['precio_estimado']))
-                $respuesta['errores']['precio_estimado'] = 'Precio estimado vacío o no es un número válido';
-        
-            
             if(count($respuesta['errores']) == 0){
         
                 return $respuesta['success'] = true;
@@ -189,43 +199,57 @@ class Actividades_Model extends Model{
         }
     }
 
-    // function update($usuario)
-    // {
-    //     try {
-    //         $sql = "UPDATE usuarios SET usuario = '{$usuario['usuario']}', clave = '{$usuario['clave']}', email = '{$usuario['email']}' WHERE usuario = '{$usuario['usuario']}'";
-    //         $result =  $this->db->query($sql);
+    function getActividadUsuario($idactividad, $idusuario)
+    {
+        // Limpiamos las variables para su uso en la consulta SQL
+        $idactividad = $this->db->real_escape_string($idactividad);
+        $idusuario = $this->db->real_escape_string($idusuario);
 
-	// 			if ($result) {
-    //                 //Warning si falla la subida de la foto, falta implementar
-	// 				if (isset($warning))
-	// 					$response['success'] = msg_success_editUser_function . "<br />" . $warning;
-	// 				else
-	// 					$response['success'] = msg_success_editUser_function;
-	// 				return $response;
-	// 			} else {
-	// 				$response['error'] = msg_error_editUser_function;
-	// 				return $response;
-	// 			}
+        try
+        {
+            $sql = "SELECT * FROM actividades WHERE idactividad = '$idactividad' AND idusuario = '$idusuario'";
+            $query =  $this->db->query($sql);
+            
+            if ($query->num_rows == 1)
+                $resultado = $query->fetch_object();
+                
+            $query->close();
+                
+            if (isset($resultado))
+                return $resultado;
+            else
+                return NULL;
+        }
+        catch(Exception $e)
+        {
+            echo "<div class='info-error'>Error:".$e->getMessage()."</div>";
+        }
+    }
 
-    //     } catch(Exception $e){
-    //         echo 'ERROR:'.$e->getMessage()."<br>";
-    //         die();
-    //     }
-    // }
 
-    // function delete($usuario)
-    // {
-    //     try
-    //     {
-    //         $sql = "DELETE FROM usuarios WHERE usuario = '{$usuario['usuario']}'";
-    //         $query = $this->db->query($sql);
+    /*
+        Dado un id de actividad y un idusuario borrará la actividad de dicho usuario
+        de la tabla actividades de la BD, al borrar la actividad borrará en cascada
+        las filas con referencias a esta actividad que tenga en las otras tablas (tareas y registros)
+     */
+    function borrarActividadUsuario($idactividad, $idusuario)
+    {
+        // Limpiamos los parametros de entrada
+        $idactividad = $this->db->real_escape_string($idactividad);
+        $idusuario = $this->db->real_escape_string($idusuario);
 
-    //         if ($query)
-    //             return "Ha borrado bien el usuario";
+        try
+        {
+            $sql = "DELETE FROM actividades WHERE idactividad = '{$idactividad}' AND idusuario = '{$idusuario}'";
+            
+            if ($this->db->query($sql))
+                return true;
+            else
+                return false;
 
-    //     } catch(Exception $e){
-    //         echo 'ERROR:'.$e->getMessage()."<br>";
-    //         die();
-    //     }
-    // }
+        } catch(Exception $e){
+            echo 'ERROR:'.$e->getMessage()."<br>";
+            die();
+        }
+    }
 }