+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Todos los cambios importantes de este proyecto serán documentados en este fiche
 
 ## [unreleased]
 
+## [Timetracker v0.9.0 (2021-11-18)][0.9.0]
+### Added
+    - Añadida la página de Crear Tarea
+    - Añadida la página de Editar Actividad
+    - Añadida la página de Editar Tarea
+    
+### Changed
+    - Modificada Crear Actividad
+
 ## [Timetracker v0.8.0 (2021-11-18)][0.8.0]
 ### Added
     - Añadida la página de Informes y Ver informe (informes.php y registroTiempo_model.php)
@@ -87,6 +96,7 @@ Todos los cambios importantes de este proyecto serán documentados en este fiche
 
 
 [unreleased]: https://github.com/eorhed/timetracker/projects/1
+[0.8.0]: https://github.com/eorhed/timetracker/commit/b552a5b19c6947ab3909b3523e9b8ae353e793cd
 [0.7.0]: https://github.com/eorhed/timetracker/commit/1446a84691553ae9bc82f2a20e3b6be25c78dd11
 [0.6.2]: https://github.com/eorhed/timetracker/commit/74c18975425072ce367dc9759e97ef69bdbb2254
 [0.6.1]: https://github.com/eorhed/timetracker/commit/1ef310be35639799f045cfd86341ea6a0a50fc3d